FoodStickersMessages/requirements.txt
Flask>=0.8
twilio>=3.3.6
pillow>=4.0.0
requests>=2.13.0